+++ packages/d2ts/src/d2ql/index.ts
@@ -8,3 +8,15 @@ export * from './schema.js'
 
 // Export the compiler functions
 export { compileQuery } from './compiler.js'
+
+// Export utility functions for external use
+export * from './utils.js'
+
+// Export extractors for external use
+export * from './extractors.js'
+
+// Export evaluators for external use
+export * from './evaluators.js'
+
+// Export join processing functions for external use
+export * from './joins.js'
